requirements.txt
paho-mqtt==2.0.0
python-decouple==3.8
